requirements.txt
opencv-python>=4.5.4
numpy>=1.20.0
PyYAML>=6.0
ultralytics>=8.0.0
torch>=1.7.0
torchvision>=0.8.0
scipy>=1.7.0
matplotlib>=3.5.0
# Deep SORT dependencies
easydict>=1.9
scikit-image>=0.18.0 